#1125 Move observability sample to correct package
