+++ hartshorn-launchpad-samples/application-observability/src/main/java/org/dockbox/sample/observability/SampleApplicationObserver.java
@@ -1,4 +1,4 @@
-package org.dockbox.sample.java;
+package org.dockbox.sample.observability;
 
 import org.dockbox.hartshorn.inject.annotations.Component;
 import org.dockbox.hartshorn.inject.annotations.Inject;
